Creating a file for option and helper functions
